requirements.txt
alembic==1.15.2
aniso8601==10.0.1
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.4
cssselect2==0.8.0
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Dance==7.1.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-REST==1.3
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.2
greenlet==3.1.1
h11==0.16.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
lxml==5.4.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ms-identity-python @ git+https://github.com/azure-samples/ms-identity-python@834583646df315b79bc89f54808f4ae9c2a1ce59
msal==1.35.1
msgspec==0.20.0
mysqlclient==2.2.7
oauthlib==3.3.1
ordered-set==4.1.0
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.11.0
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dotenv==1.1.0
python-engineio==4.13.0
python-slugify==8.0.4
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.1
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.5.0
tinyhtml5==2.0.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
URLObject==3.0.0
weasyprint==68.0
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==1.17.2
wsproto==1.3.2
WTForms==3.2.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1

auth0-server-python>=1.0.0b7
flask[async]>=2.0.0